// src/app/shared/models/Manager/update-vehicle-with-spec.model.ts
export interface UpdateVehicleWithSpec {
  vehicleName: string;
  model: string;
  yearOfProduction: number;
  basePrice: number;
  stockCount: number;
  shortDescription: string;

  engine: number;
  powerOfVehical: string;
  torque: string;
  fuelType: string;
  mileage: string;
  bodyType: string;
  seatingCapacity: number;
}